negotiate-ai-basic/negotiate-ai/requirements.txt
streamlit==1.41.0
plotly==5.24.0
google-genai==1.14.0
elevenlabs==2.0.0
pymongo==4.12.1
solders==0.21.0
requests==2.32.0
